CRSTemplate: Move parameter implementations to separate files and namespace
