+++ src/global.cpp
@@ -23,46 +23,13 @@
 
 #include <mapper_config.h>
 
-#include "core/crs_template.h"
+#include "core/crs_template_implementation.h"
 #include "core/georeferencing.h"
 #include "file_format_registry.h"
 #include "file_format_native.h"
 #include "file_format_ocad8.h"
 #include "file_format_xml.h"
 #include "fileformats/ocd_file_format.h"
-
-void registerProjectionTemplates()
-{
-	// TODO: These objects are never deleted.
-	
-	/*
-	 * CRSTemplate(
-	 *     id,
-	 *     name,
-	 *     coordinates name,
-	 *     specification string)
-	 * 
-	 * The id must be unique and different from "Local".
-	 */
-	
-	// UTM
-	CRSTemplate* temp = new CRSTemplate(
-		"UTM",
-		Georeferencing::tr("UTM", "UTM coordinate reference system"),
-		Georeferencing::tr("UTM coordinates"),
-		"+proj=utm +datum=WGS84 +zone=%1");
-	temp->addParam(new CRSTemplate::UTMZoneParam(Georeferencing::tr("UTM Zone (number north/south, e.g. \"32 N\", \"24 S\")")));
-	CRSTemplate::registerCRSTemplate(temp);
-	
-	// Gauss-Krueger
-	temp = new CRSTemplate(
-		"Gauss-Krueger, datum: Potsdam",
-		Georeferencing::tr("Gauss-Krueger, datum: Potsdam", "Gauss-Krueger coordinate reference system"),
-		Georeferencing::tr("Gauss-Krueger coordinates"),
-		"+proj=tmerc +lat_0=0 +lon_0=%1 +k=1.000000 +x_0=%2 +y_0=0 +ellps=bessel +datum=potsdam +units=m +no_defs");
-	temp->addParam((new CRSTemplate::IntRangeParam(Georeferencing::tr("Zone number (1 to 119)", "Zone number for Gauss-Krueger coordinates"), 1, 119))->clearOutputs()->addDerivedOutput(3, 0)->addDerivedOutput(1000000, 500000));
-	CRSTemplate::registerCRSTemplate(temp);
-}
 
 void doStaticInitializations()
 {
@@ -79,5 +46,5 @@ void doStaticInitializations()
 #endif
 	
 	// Register projection templates
-	registerProjectionTemplates();
+	CRSTemplates::registerProjectionTemplates();
 }
