+++ test/georeferencing_t.cpp
@@ -21,6 +21,7 @@
 
 #include <proj_api.h>
 
+#include "../src/core/crs_template.h"
 #include "../src/file_format_xml.h"
